requirements.txt
lxml
pyparsing
python-dateutil
requests
schematics
